--- src/uttaksplan/utils/uttaksgrunnlagUtils.ts
@@ -1,8 +1,0 @@
-import { UttaksplanRequiredProps } from 'uttaksplan/uttak/types';
-
-export function getErDeltPermisjon(props: UttaksplanRequiredProps): boolean {
-    if (!props.annenForelder) {
-        return false;
-    }
-    return true;
-}
